accessible/tests/mochitest/test_table_2.html

<!DOCTYPE HTML PUBLIC "-//w3c//dtd html 4.0 transitional//en">
<html>
 <head>
  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
  <title> nsIAccessibleTable Interface Test Case </title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>

  <script type="text/javascript">

function doTest()
{
  // bug 410052
  var accTable = getAccessible("table", [nsIAccessibleTable]);

  is(accTable.getIndexAt(2,4), 17, "wrong index returned");
  is(accTable.getColumnExtentAt(2,2), 2, "colspan wrong");
  is(accTable.getColumnExtentAt(0,3), 2, "colspan wrong");
  is(accTable.getColumnExtentAt(3,5), 1, "colspan");
  is(accTable.getRowExtentAt(0,7), 4, "rowspan wrong");
  is(accTable.getRowExtentAt(2,7), 4,"rowspan wrong");
  is(accTable.getColumnExtentAt(2,3), 1, "colspan wrong");
  is(accTable.cellRefAt(2,1).firstChild.name, "c1", "wrong cell");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doTest);
  </script>
 </head>
 <body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=410052">Mozilla Bug 410052</a>
  <br>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=417912"
                     title="GetCellDataAt callers that expect an error if no cell is found are wrong">
    Mozilla Bug 417912
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <!-- Test Table -->
  <br><br><b> Testing Table:</b><br><br>
  <center>
   <table cellpadding="2" cellspacing="2" border="1" width="50%" id="table">
    <tbody>
     <tr>
      <td><br></td>
      <td><br></td>
      <td rowspan="1" colspan="2"><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td rowspan="4" colspan="1"><br></td>
     </tr>
     <tr>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
     </tr>
     <tr>
      <td><br></td>
      <td rowspan="2" colspan="2">c1</td>
      <td><br></td>
      <td><br></td>
      <td rowspan="2" colspan="1"><br></td>
      <td><br></td>
     </tr>
     <tr>
      <td><br></td>
      <td><br></td>
      <td><br></td>
      <td><br></td>
     </tr>
    </tbody>
   </table>
  </center>
 </body>
</html>
